refactor(compute): move remaining ops into kernel modules
