+++ crates/compute/src/kernels/detect_contacts_sdf_op.rs
@@ -1,0 +1,9 @@
+use crate::{BufferView, ComputeError};
+
+pub fn handle_detect_contacts_sdf(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, ComputeError> {
+    if !binds.is_empty() {
+        Ok(vec![binds[0].data.to_vec()])
+    } else {
+        Ok(Vec::new())
+    }
+}
