+++ crates/compute/src/kernels/mod.rs
@@ -11,4 +11,22 @@ pub mod max_op;
 pub mod clamp_op;
 pub mod exp_op;
 pub mod log_op;
-// Add other kernel modules here as they are created. 
+pub mod sqrt_op;
+pub mod rsqrt_op;
+pub mod tanh_op;
+pub mod relu_op;
+pub mod sigmoid_op;
+pub mod reduce_sum_op;
+pub mod reduce_mean_op;
+pub mod reduce_max_op;
+pub mod segmented_reduce_sum_op;
+pub mod scatter_add_op;
+pub mod gather_op;
+pub mod matmul_op;
+pub mod integrate_bodies_op;
+pub mod detect_contacts_sdf_op;
+pub mod solve_contacts_pbd_op;
+pub mod solve_joints_pbd_op;
+pub mod rng_normal_op;
+pub mod expand_instances_op;
+// Add other kernel modules here as they are created.
