+++ crates/compute/src/kernels/sigmoid_op.rs
@@ -1,15 +1,15 @@
 use crate::{BufferView, ComputeError};
 
-pub fn handle_sqrt(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, ComputeError> {
+pub fn handle_sigmoid(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, ComputeError> {
     if binds.len() < 3 {
-        return Err(ComputeError::ShapeMismatch("Sqrt kernel expects 3 buffers"));
+        return Err(ComputeError::ShapeMismatch("Sigmoid kernel expects 3 buffers"));
     }
     let input_view = &binds[0];
     if input_view.element_size_in_bytes != std::mem::size_of::<f32>() {
-        return Err(ComputeError::ShapeMismatch("Sqrt kernel currently only supports f32 data"));
+        return Err(ComputeError::ShapeMismatch("Sigmoid kernel currently only supports f32 data"));
     }
     let input_values: &[f32] = bytemuck::cast_slice(&input_view.data);
-    let output_values: Vec<f32> = input_values.iter().map(|&x| x.sqrt()).collect();
+    let output_values: Vec<f32> = input_values.iter().map(|&x| 1.0 / (1.0 + (-x).exp())).collect();
     let out_bytes = bytemuck::cast_slice(&output_values).to_vec();
     Ok(vec![out_bytes])
 }
@@ -21,10 +21,10 @@ mod tests {
     use std::sync::Arc as StdArc;
 
     #[test]
-    fn mock_sqrt_computes_square_root() {
+    fn mock_sigmoid_computes_logistic_function() {
         let cpu = MockCpu::default();
-        let input_data = vec![0.0f32, 1.0, 4.0, 9.0, 2.0];
-        let expected_output_data: Vec<f32> = input_data.iter().map(|&x| x.sqrt()).collect();
+        let input_data = vec![0.0f32, 1.0, -1.0, 0.5, -0.5, 20.0, -20.0];
+        let expected_output_data: Vec<f32> = input_data.iter().map(|&x| 1.0 / (1.0 + (-x).exp())).collect();
 
         let input_bytes: StdArc<[u8]> = bytemuck::cast_slice(&input_data).to_vec().into();
         let input_buffer_view = BufferView::new(input_bytes, vec![input_data.len()], std::mem::size_of::<f32>());
@@ -37,7 +37,7 @@ mod tests {
         let config_buffer_view = BufferView::new(config_bytes, vec![config_data.len()], std::mem::size_of::<u32>());
 
         let dispatch_binds = [input_buffer_view, output_buffer_view, config_buffer_view];
-        let result_buffers = cpu.dispatch(&Kernel::Sqrt, &dispatch_binds, [1,1,1]).expect("Dispatch for Sqrt failed");
+        let result_buffers = cpu.dispatch(&Kernel::Sigmoid, &dispatch_binds, [1,1,1]).expect("Dispatch for Sigmoid failed");
 
         assert_eq!(result_buffers.len(), 1);
         let output_bytes = &result_buffers[0];
@@ -45,7 +45,7 @@ mod tests {
         let output_values: &[f32] = bytemuck::cast_slice(output_bytes);
         assert_eq!(output_values.len(), expected_output_data.len());
         for (got, expected) in output_values.iter().zip(expected_output_data.iter()) {
-            assert!((got - expected).abs() < 1e-6, "Mismatch for Sqrt. Got: {}, Expected: {}", got, expected);
+            assert!((got - expected).abs() < 1e-6, "Mismatch for Sigmoid. Got: {}, Expected: {}", got, expected);
         }
     }
 }
