+++ crates/compute/src/kernels/solve_joints_pbd_op.rs
@@ -1,0 +1,9 @@
+use crate::{BufferView, ComputeError};
+
+pub fn handle_solve_joints_pbd(binds: &[BufferView]) -> Result<Vec<Vec<u8>>, ComputeError> {
+    if !binds.is_empty() {
+        Ok(vec![binds[0].data.to_vec()])
+    } else {
+        Ok(Vec::new())
+    }
+}
